Move the PersonalInformation.jsx to a separate 'personalInformation' folder.
