+++ src/components/form/components/personalInformation/ui/PersonalInformation.jsx
@@ -1,5 +1,5 @@
 import React, {useContext} from "react";
-import {ResumeContext} from "../../builder";
+import {ResumeContext} from "../../../../builder";
 
 const PersonalInformation = ({}) => {
   const {resumeData, setResumeData, handleProfilePicture, handleChange} =
